[CHORE] Eclipse Project transfer to Intelij

+++ workspace/ch03/src/OperatorEx06.java
@@ -2,7 +2,7 @@ class OperatorEx06 {
       public static void main(String[] args) { 
             byte a = 10; 
             byte b = 20; 
-            byte c = a + b; 
+            byte c = (byte)(a + b);
 
             System.out.println(c); 
       }
